requirements.txt
selenium==3.141.0
tinydb==4.1.1
pyyaml==5.3.1
pyTelegramBotAPI==3.7.1
